+++ src/main/resources/templates/order/deliveryform.html
@@ -95,7 +95,7 @@
                     <span>기본배송지로 저장</span>
                 </div>
                 <div class="d-flex justify-content-center">
-                    <a href="javascript:window.history.back();" class="btn btn-outline-dark w-75 me-3">취소</a>
+                    <a th:href="@{list}" class="btn btn-outline-dark w-75 me-3">취소</a>
                     <button type="submit" class="btn btn-primary w-75">저장</button>
                 </div>
             </form>
